Mistral_OCR_app/OCR_app/requirements.txt
flask>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
